responces

+++ src/app/pages/admin/auditories/auditories.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { ConfirmationService, MessageService } from 'primeng/api';
+import { skipWhile, take } from 'rxjs';
 import { Auditory } from '../../../api/interfaces/auditory.interface';
 import { AuditoryService } from '../services/auditory.service';
 
@@ -18,7 +20,16 @@ export class AdminAuditoriesComponent {
   });
   auditories: Auditory[] = [];
 
-  constructor(private readonly auditoryService: AuditoryService) {
+  deleteAuditoryResponce$ = this.auditoryService.deleteAuditoryResponce$;
+  createAuditoryResponce$ = this.auditoryService.createAuditoryResponce$;
+  updateAuditoryResponce$ = this.auditoryService.updateAuditoryResponce$;
+  getAllAuditoriesResponce$ = this.auditoryService.getAllAuditoriesResponce$;
+
+  constructor(
+    private readonly auditoryService: AuditoryService,
+    private messageService: MessageService,
+    private confirmationService: ConfirmationService
+  ) {
     auditoryService.auditories$.subscribe((auditories) => {
       this.auditories = auditories;
     });
@@ -46,14 +57,43 @@ export class AdminAuditoriesComponent {
     const { auditoryName } = this.auditoryForm.value;
     if (this.isEditing) {
       this.auditoryService.updateAuditory(this.auditoryToEditId, auditoryName!);
+      this.updateAuditoryResponce$
+        .pipe(
+          skipWhile((responce) => responce === 'pending'),
+          take(1)
+        )
+        .subscribe((responce) => {
+          console.log(responce);
+          if (responce === 'success') {
+            this.isFormAuditoryOpen = false;
+            this.messageService.add({
+              severity: 'success',
+              summary: 'Данные об аудитории обновлены',
+            });
+          }
+        });
     } else {
       this.auditoryService.createAuditory(auditoryName!);
+      this.createAuditoryResponce$
+        .pipe(
+          skipWhile((responce) => responce === 'pending'),
+          take(1)
+        )
+        .subscribe((responce) => {
+          if (responce === 'success') {
+            this.isFormAuditoryOpen = false;
+            this.messageService.add({
+              severity: 'success',
+              summary: 'Данные об аудитории добавлены',
+            });
+          }
+        });
     }
   }
 
   editAuditory(auditory: Auditory) {
     this.auditoryToEditId = auditory.id;
-    this.isEditing = false;
+    this.isEditing = true;
     this.openFormDialog();
     this.auditoryForm.patchValue({
       auditoryName: auditory.name,
@@ -61,6 +101,25 @@ export class AdminAuditoriesComponent {
   }
 
   deleteAuditory(auditory: Auditory) {
-    this.auditoryService.deleteAuditory(auditory);
+    this.confirmationService.confirm({
+      message: 'Вы действительно хотите данные об аудитории?',
+      accept: () => {
+        this.auditoryService.deleteAuditory(auditory);
+
+        this.deleteAuditoryResponce$
+          .pipe(
+            skipWhile((responce) => responce === 'pending'),
+            take(1)
+          )
+          .subscribe((responce) => {
+            if (responce === 'success') {
+              this.messageService.add({
+                severity: 'success',
+                summary: 'Данные об аудитории удалены',
+              });
+            }
+          });
+      },
+    });
   }
 }
