+++ src/app/pages/admin/event-tags/event-tags.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { ConfirmationService, MessageService } from 'primeng/api';
+import { skipWhile, take } from 'rxjs';
 import { EventTag } from 'src/app/api/interfaces/eventTag.interface';
 import { EventTagService } from 'src/app/pages/admin/services/eventTag.service';
 import { Auditory } from '../../../api/interfaces/auditory.interface';
@@ -21,7 +23,16 @@ export class AdminEventTagsComponent {
 
   eventTags: EventTag[] = [];
 
-  constructor(private readonly eventTagService: EventTagService) {
+  deleteEventTagResponce$ = this.eventTagService.deleteEventTagResponce$;
+  createEventTagResponce$ = this.eventTagService.createEventTagResponce$;
+  updateEventTagResponce$ = this.eventTagService.updateEventTagResponce$;
+  getAllEventsTagResponce$ = this.eventTagService.getAllEventsTagResponce$;
+
+  constructor(
+    private readonly eventTagService: EventTagService,
+    private messageService: MessageService,
+    private confirmationService: ConfirmationService
+  ) {
     eventTagService.eventTags$.subscribe((auditories) => {
       this.eventTags = auditories;
     });
@@ -49,8 +60,37 @@ export class AdminEventTagsComponent {
     if (this.isEditing) {
       this.eventTagService.updateEventTag(this.eventTagIdEdit, eventTagName!);
       this.isEditing = false;
+      this.updateEventTagResponce$
+        .pipe(
+          skipWhile((responce) => responce === 'pending'),
+          take(1)
+        )
+        .subscribe((responce) => {
+          console.log(responce);
+          if (responce === 'success') {
+            this.isFormDialogOpen = false;
+            this.messageService.add({
+              severity: 'success',
+              summary: 'Данные об теге мероприятия обновлены',
+            });
+          }
+        });
     } else {
       this.eventTagService.createEventTag(eventTagName!);
+      this.createEventTagResponce$
+        .pipe(
+          skipWhile((responce) => responce === 'pending'),
+          take(1)
+        )
+        .subscribe((responce) => {
+          if (responce === 'success') {
+            this.isFormDialogOpen = false;
+            this.messageService.add({
+              severity: 'success',
+              summary: 'Данные об теге мероприятия добавлены',
+            });
+          }
+        });
     }
   }
 
@@ -61,6 +101,25 @@ export class AdminEventTagsComponent {
     this.eventTagForm.patchValue({ eventTagName: eventTag.name });
   }
   deleteEventTag(group: EventTag) {
-    this.eventTagService.deleteEventTag(group);
+    this.confirmationService.confirm({
+      message: 'Вы действительно хотите данные о теге мероприятия?',
+      accept: () => {
+        this.eventTagService.deleteEventTag(group);
+
+        this.deleteEventTagResponce$
+          .pipe(
+            skipWhile((responce) => responce === 'pending'),
+            take(1)
+          )
+          .subscribe((responce) => {
+            if (responce === 'success') {
+              this.messageService.add({
+                severity: 'success',
+                summary: 'Данные о теге мероприятия удалены',
+              });
+            }
+          });
+      },
+    });
   }
 }
