+++ src/app/pages/admin/groups/groups.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { ConfirmationService, MessageService } from 'primeng/api';
+import { skipWhile, take } from 'rxjs';
 import { Group } from 'src/app/api/interfaces/group.interface';
 import { GroupService } from 'src/app/pages/admin/services/group.service';
 
@@ -18,7 +20,17 @@ export class AdminGroupsComponent {
 
   groups: Group[] = [];
 
-  constructor(private readonly groupService: GroupService) {
+  deleteGroupResponce$ = this.groupService.deleteGroupResponce$;
+  createGroupResponce$ = this.groupService.createGroupResponce$;
+  updateGroupResponce$ = this.groupService.updateGroupResponce$;
+  getAllGroupsResponce$ = this.groupService.getAllGroupsResponce$;
+
+  constructor(
+    private readonly groupService: GroupService,
+
+    private messageService: MessageService,
+    private confirmationService: ConfirmationService
+  ) {
     groupService.groups$.subscribe((groups) => {
       this.groups = groups;
     });
@@ -31,7 +43,26 @@ export class AdminGroupsComponent {
     this.openAddDialog();
   }
   deleteGroup(group: Group) {
-    this.groupService.deleteGroup(group);
+    this.confirmationService.confirm({
+      message: 'Вы действительно хотите данные о группе?',
+      accept: () => {
+        this.groupService.deleteGroup(group);
+
+        this.deleteGroupResponce$
+          .pipe(
+            skipWhile((responce) => responce === 'pending'),
+            take(1)
+          )
+          .subscribe((responce) => {
+            if (responce === 'success') {
+              this.messageService.add({
+                severity: 'success',
+                summary: 'Данные о группе удалены',
+              });
+            }
+          });
+      },
+    });
   }
 
   openAddDialog() {
@@ -50,8 +81,37 @@ export class AdminGroupsComponent {
     const { groupName } = this.groupForm.value;
     if (this.isEditing) {
       this.groupService.updateGroup(this.groupIdEdit, groupName!);
+      this.updateGroupResponce$
+        .pipe(
+          skipWhile((responce) => responce === 'pending'),
+          take(1)
+        )
+        .subscribe((responce) => {
+          console.log(responce);
+          if (responce === 'success') {
+            this.isformGroupDialogOpen = false;
+            this.messageService.add({
+              severity: 'success',
+              summary: 'Данные о группе обновлены',
+            });
+          }
+        });
     } else {
       this.groupService.createGroup(groupName!);
+      this.createGroupResponce$
+        .pipe(
+          skipWhile((responce) => responce === 'pending'),
+          take(1)
+        )
+        .subscribe((responce) => {
+          if (responce === 'success') {
+            this.isformGroupDialogOpen = false;
+            this.messageService.add({
+              severity: 'success',
+              summary: 'Данные о группе добавлены',
+            });
+          }
+        });
     }
   }
 }
